requirements.txt
Flask==2.0.1
line-bot-sdk==2.0.1
gunicorn==20.1.0
